Removed unused contracts and setup testing scaffold

+++ contracts/interfaces/IwsUMAMI.sol
@@ -5,6 +5,7 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IwsUMAMI is IERC20 {
-  function wrap( uint _amount ) external returns ( uint );
-  function unwrap( uint _amount ) external returns ( uint );
+    function wrap(uint256 _amount) external returns (uint256);
+
+    function unwrap(uint256 _amount) external returns (uint256);
 }
